+++ sdk/resourcemanager/privatedns/armprivatedns/zz_generated_virtualnetworklinks_client.go
@@ -12,15 +12,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"net/url"
-	"strconv"
-	"strings"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
+	"net/url"
+	"strconv"
+	"strings"
 )
 
 // VirtualNetworkLinksClient contains the methods for the VirtualNetworkLinks group.
@@ -32,8 +32,15 @@ type VirtualNetworkLinksClient struct {
 }
 
 // NewVirtualNetworkLinksClient creates a new instance of VirtualNetworkLinksClient with the specified values.
-func NewVirtualNetworkLinksClient(con *arm.Connection, subscriptionID string) *VirtualNetworkLinksClient {
-	return &VirtualNetworkLinksClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version), subscriptionID: subscriptionID}
+func NewVirtualNetworkLinksClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) *VirtualNetworkLinksClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &VirtualNetworkLinksClient{subscriptionID: subscriptionID, ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // BeginCreateOrUpdate - Creates or updates a virtual network link to the specified Private DNS zone.
@@ -260,7 +267,7 @@ func (client *VirtualNetworkLinksClient) getCreateRequest(ctx context.Context, r
 func (client *VirtualNetworkLinksClient) getHandleResponse(resp *http.Response) (VirtualNetworkLinksGetResponse, error) {
 	result := VirtualNetworkLinksGetResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.VirtualNetworkLink); err != nil {
-		return VirtualNetworkLinksGetResponse{}, err
+		return VirtualNetworkLinksGetResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }
@@ -325,7 +332,7 @@ func (client *VirtualNetworkLinksClient) listCreateRequest(ctx context.Context, 
 func (client *VirtualNetworkLinksClient) listHandleResponse(resp *http.Response) (VirtualNetworkLinksListResponse, error) {
 	result := VirtualNetworkLinksListResponse{RawResponse: resp}
 	if err := runtime.UnmarshalAsJSON(resp, &result.VirtualNetworkLinkListResult); err != nil {
-		return VirtualNetworkLinksListResponse{}, err
+		return VirtualNetworkLinksListResponse{}, runtime.NewResponseError(err, resp)
 	}
 	return result, nil
 }